global formatting code, hide api keys to keystore.properties, refactoring
