+++ app/src/main/res/anim/fade_in.xml
@@ -4,5 +4,5 @@
         android:duration="@integer/anim_fade_duration"
         android:fillAfter="true"
         android:fromAlpha="0.0"
-        android:toAlpha="1.0"/>
+        android:toAlpha="1.0" />
 </set>
