+++ app/src/main/res/anim/fade_out.xml
@@ -4,5 +4,5 @@
         android:duration="@integer/anim_fade_duration"
         android:fillAfter="true"
         android:fromAlpha="1.0"
-        android:toAlpha="0.0"/>
+        android:toAlpha="0.0" />
 </set>
