+++ app/src/main/res/drawable/img_clouds.xml
@@ -1,8 +1,21 @@
-<vector android:height="24dp" android:viewportHeight="481.306"
-    android:viewportWidth="481.306" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FCB641" android:pathData="M205.6,233.2a136,133.5 0,1 0,272 0a136,133.5 0,1 0,-272 0z"/>
-    <path android:fillColor="#FCB641" android:pathData="M336,65.5c-4.4,0 -8,-3.6 -8,-8v-48c0,-4.4 3.6,-8 8,-8s8,3.6 8,8v48C344,61.9 340.4,65.5 336,65.5z"/>
-    <path android:fillColor="#FCB641" android:pathData="M439.4,98.8c-2,0 -4.1,-0.8 -5.7,-2.3c-3.1,-3.1 -3.1,-8.2 0,-11.3l33.9,-33.9c3.1,-3.1 8.2,-3.1 11.3,0s3.1,8.2 0,11.3l-33.9,33.9C443.5,98 441.4,98.8 439.4,98.8z"/>
-    <path android:fillColor="#FCB641" android:pathData="M209,122.4c-2,0 -4.1,-0.8 -5.7,-2.3l-33.9,-33.9c-3.1,-3.1 -3.1,-8.2 0,-11.3s8.2,-3.1 11.3,0l33.9,33.9c3.1,3.1 3.1,8.2 0,11.3C213.1,121.6 211,122.4 209,122.4z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M426.2,361.5c-1.1,0 -42.6,0 -42.6,0c0.7,-8 1.2,-8.7 1.3,-13.5c0.1,-42.7 -28.4,-80 -70,-89.1c-23.8,-5.2 -46.4,-1.4 -65.2,8.8c-22.9,-34.1 -60.8,-57 -104.3,-59.7c-5.6,-0.8 -11.3,-1.2 -17.1,-1.2C57.4,206.7 0,265.9 0,339c0,38.1 15.7,72.3 40.6,96.4c22.1,24.5 52.7,40.3 89,42.9c0,0 295.4,1.5 296.6,1.5c29.7,0 53.8,-27.2 53.8,-59.2C480,388.7 455.9,361.5 426.2,361.5z"/>
+<vector android:height="24dp"
+    android:viewportHeight="481.306"
+    android:viewportWidth="481.306"
+    android:width="24dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#FCB641"
+        android:pathData="M205.6,233.2a136,133.5 0,1 0,272 0a136,133.5 0,1 0,-272 0z" />
+    <path
+        android:fillColor="#FCB641"
+        android:pathData="M336,65.5c-4.4,0 -8,-3.6 -8,-8v-48c0,-4.4 3.6,-8 8,-8s8,3.6 8,8v48C344,61.9 340.4,65.5 336,65.5z" />
+    <path
+        android:fillColor="#FCB641"
+        android:pathData="M439.4,98.8c-2,0 -4.1,-0.8 -5.7,-2.3c-3.1,-3.1 -3.1,-8.2 0,-11.3l33.9,-33.9c3.1,-3.1 8.2,-3.1 11.3,0s3.1,8.2 0,11.3l-33.9,33.9C443.5,98 441.4,98.8 439.4,98.8z" />
+    <path
+        android:fillColor="#FCB641"
+        android:pathData="M209,122.4c-2,0 -4.1,-0.8 -5.7,-2.3l-33.9,-33.9c-3.1,-3.1 -3.1,-8.2 0,-11.3s8.2,-3.1 11.3,0l33.9,33.9c3.1,3.1 3.1,8.2 0,11.3C213.1,121.6 211,122.4 209,122.4z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M426.2,361.5c-1.1,0 -42.6,0 -42.6,0c0.7,-8 1.2,-8.7 1.3,-13.5c0.1,-42.7 -28.4,-80 -70,-89.1c-23.8,-5.2 -46.4,-1.4 -65.2,8.8c-22.9,-34.1 -60.8,-57 -104.3,-59.7c-5.6,-0.8 -11.3,-1.2 -17.1,-1.2C57.4,206.7 0,265.9 0,339c0,38.1 15.7,72.3 40.6,96.4c22.1,24.5 52.7,40.3 89,42.9c0,0 295.4,1.5 296.6,1.5c29.7,0 53.8,-27.2 53.8,-59.2C480,388.7 455.9,361.5 426.2,361.5z" />
 </vector>
