+++ app/src/main/res/drawable/img_clouds_night.xml
@@ -1,7 +1,18 @@
-<vector android:height="24dp" android:viewportHeight="480.016"
-    android:viewportWidth="480.016" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#92C7D3" android:pathData="M426.2,363.8c-1.1,0 -42.6,0 -42.6,0c0.7,-8.2 1.2,-10 1.3,-14.7c0.1,-42.7 -28.4,-80.6 -70,-89.8c-23.8,-5.2 -46.4,-1.7 -65.2,8.4c-22.9,-34.1 -60.8,-57.2 -104.3,-59.9c-5.6,-0.8 -11.3,-1.3 -17.1,-1.3C57.4,206.6 0,265.8 0,338.9c0,38.1 15.7,72.3 40.6,96.4c22.1,24.5 52.7,40.3 89,42.9c0,0 295.4,1.5 296.6,1.5c29.7,0 53.8,-25.9 53.8,-57.9C480,389.8 455.9,363.8 426.2,363.8z"/>
-    <path android:fillColor="#FCE674" android:pathData="M363,2c10.3,17.2 16.3,37.3 16.3,58.8c0,63.4 -51.4,114.9 -114.9,114.9c-0.1,0 -0.3,-0 -0.4,-0c20.1,33.6 56.6,56.1 98.6,56.1c63.4,0 114.9,-51.4 114.9,-114.9C477.5,53.5 426.2,2.2 363,2z"/>
-    <path android:fillColor="#FCE674" android:pathData="M76.6,149.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z"/>
-    <path android:fillColor="#FCE674" android:pathData="M192.4,0.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z"/>
+<vector android:height="24dp"
+    android:viewportHeight="480.016"
+    android:viewportWidth="480.016"
+    android:width="24dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M426.2,363.8c-1.1,0 -42.6,0 -42.6,0c0.7,-8.2 1.2,-10 1.3,-14.7c0.1,-42.7 -28.4,-80.6 -70,-89.8c-23.8,-5.2 -46.4,-1.7 -65.2,8.4c-22.9,-34.1 -60.8,-57.2 -104.3,-59.9c-5.6,-0.8 -11.3,-1.3 -17.1,-1.3C57.4,206.6 0,265.8 0,338.9c0,38.1 15.7,72.3 40.6,96.4c22.1,24.5 52.7,40.3 89,42.9c0,0 295.4,1.5 296.6,1.5c29.7,0 53.8,-25.9 53.8,-57.9C480,389.8 455.9,363.8 426.2,363.8z" />
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M363,2c10.3,17.2 16.3,37.3 16.3,58.8c0,63.4 -51.4,114.9 -114.9,114.9c-0.1,0 -0.3,-0 -0.4,-0c20.1,33.6 56.6,56.1 98.6,56.1c63.4,0 114.9,-51.4 114.9,-114.9C477.5,53.5 426.2,2.2 363,2z" />
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M76.6,149.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z" />
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M192.4,0.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z" />
 </vector>
