+++ app/src/main/res/drawable/img_night.xml
@@ -1,6 +1,15 @@
-<vector android:height="24dp" android:viewportHeight="480.016"
-    android:viewportWidth="480.016" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FCE674" android:pathData="M363,2c10.3,17.2 16.3,37.3 16.3,58.8c0,63.4 -51.4,114.9 -114.9,114.9c-0.1,0 -0.3,-0 -0.4,-0c20.1,33.6 56.6,56.1 98.6,56.1c63.4,0 114.9,-51.4 114.9,-114.9C477.5,53.5 426.2,2.2 363,2z"/>
-    <path android:fillColor="#FCE674" android:pathData="M76.6,149.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z"/>
-    <path android:fillColor="#FCE674" android:pathData="M192.4,0.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z"/>
+<vector android:height="24dp"
+    android:viewportHeight="480.016"
+    android:viewportWidth="480.016"
+    android:width="24dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M363,2c10.3,17.2 16.3,37.3 16.3,58.8c0,63.4 -51.4,114.9 -114.9,114.9c-0.1,0 -0.3,-0 -0.4,-0c20.1,33.6 56.6,56.1 98.6,56.1c63.4,0 114.9,-51.4 114.9,-114.9C477.5,53.5 426.2,2.2 363,2z" />
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M76.6,149.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z" />
+    <path
+        android:fillColor="#FCE674"
+        android:pathData="M192.4,0.3l15,30.4l33.6,4.9l-24.3,23.7l5.7,33.4l-30,-15.8l-30,15.8l5.7,-33.4l-24.3,-23.7l33.6,-4.9z" />
 </vector>
