+++ app/src/main/res/drawable/img_storm.xml
@@ -1,13 +1,36 @@
-<vector android:height="24dp" android:viewportHeight="481.12"
-    android:viewportWidth="481.12" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#77B9C8" android:pathData="M427.2,162.4c-1.1,0 -42.7,0 -42.7,0c0.7,-8 1.2,-11.2 1.3,-16c0.1,-42.8 -28.4,-81.4 -70.2,-90.6c-23.8,-5.2 -46.5,-2 -65.3,8.2C227.2,29.9 189.2,6.5 145.7,3.9c-5.6,-0.8 -11.3,-1.4 -17.1,-1.4C57.6,2.5 0,61.8 0,135c0,38.2 15.7,72.4 40.7,96.6c22.1,24.5 52.8,40.4 89.2,43c0,0 296.1,1.5 297.2,1.5c29.8,0 53.9,-24.8 53.9,-56.9C481.1,187.2 457,162.4 427.2,162.4z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M34.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C40,477.2 37.3,478.6 34.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M82.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C88,477.2 85.3,478.6 82.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M130.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C136,477.2 133.3,478.6 130.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M178.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C184,477.2 181.3,478.6 178.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M370.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C376,477.2 373.3,478.6 370.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M322.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C328,477.2 325.3,478.6 322.6,478.6z"/>
-    <path android:fillColor="#92C7D3" android:pathData="M418.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C424,477.2 421.3,478.6 418.6,478.6z"/>
-    <path android:fillColor="#FDC567" android:pathData="M88.3,369.2l25.5,-70.2l-44.9,-25.9l102.6,-113.6l-27.4,99.6l57.5,33.2z"/>
-    <path android:fillColor="#FDC567" android:pathData="M260.6,452.2l25.5,-70.2l-44.9,-25.9l102.6,-113.6l-27.4,99.6l57.5,33.2z"/>
+<vector android:height="24dp"
+    android:viewportHeight="481.12"
+    android:viewportWidth="481.12"
+    android:width="24dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#77B9C8"
+        android:pathData="M427.2,162.4c-1.1,0 -42.7,0 -42.7,0c0.7,-8 1.2,-11.2 1.3,-16c0.1,-42.8 -28.4,-81.4 -70.2,-90.6c-23.8,-5.2 -46.5,-2 -65.3,8.2C227.2,29.9 189.2,6.5 145.7,3.9c-5.6,-0.8 -11.3,-1.4 -17.1,-1.4C57.6,2.5 0,61.8 0,135c0,38.2 15.7,72.4 40.7,96.6c22.1,24.5 52.8,40.4 89.2,43c0,0 296.1,1.5 297.2,1.5c29.8,0 53.9,-24.8 53.9,-56.9C481.1,187.2 457,162.4 427.2,162.4z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M34.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C40,477.2 37.3,478.6 34.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M82.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C88,477.2 85.3,478.6 82.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M130.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C136,477.2 133.3,478.6 130.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M178.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C184,477.2 181.3,478.6 178.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M370.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C376,477.2 373.3,478.6 370.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M322.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C328,477.2 325.3,478.6 322.6,478.6z" />
+    <path
+        android:fillColor="#92C7D3"
+        android:pathData="M418.6,478.6c-1.4,0 -2.7,-0.3 -4,-1.1c-3.8,-2.2 -5.1,-7.1 -2.9,-10.9l28,-48.5c2.2,-3.8 7.1,-5.1 10.9,-2.9c3.8,2.2 5.1,7.1 2.9,10.9l-28,48.5C424,477.2 421.3,478.6 418.6,478.6z" />
+    <path
+        android:fillColor="#FDC567"
+        android:pathData="M88.3,369.2l25.5,-70.2l-44.9,-25.9l102.6,-113.6l-27.4,99.6l57.5,33.2z" />
+    <path
+        android:fillColor="#FDC567"
+        android:pathData="M260.6,452.2l25.5,-70.2l-44.9,-25.9l102.6,-113.6l-27.4,99.6l57.5,33.2z" />
 </vector>
