+++ app/src/main/res/drawable/img_sun.xml
@@ -1,5 +1,12 @@
-<vector android:height="24dp" android:viewportHeight="483.512"
-    android:viewportWidth="483.512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FCB641" android:pathData="M483.5,241.8l-54.6,49.9l22,70.6l-72.3,15.6l-16.3,72.2l-70.5,-22.5l-50.1,54.5l-50.1,-54.5l-70.5,22.5l-16.3,-72.2l-72.4,-15.6l22,-70.6l-54.6,-49.9l54.6,-49.9l-22,-70.6l72.3,-15.6l16.3,-72.2l70.5,22.5l50.1,-54.5l50.1,54.5l70.5,-22.5l16.3,72.2l72.3,15.6l-22,70.6z"/>
-    <path android:fillColor="#FDC567" android:pathData="M241.8,244.5m-138.7,0a138.7,138.7 0,1 1,277.5 0a138.7,138.7 0,1 1,-277.5 0"/>
+<vector android:height="24dp"
+    android:viewportHeight="483.512"
+    android:viewportWidth="483.512"
+    android:width="24dp"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#FCB641"
+        android:pathData="M483.5,241.8l-54.6,49.9l22,70.6l-72.3,15.6l-16.3,72.2l-70.5,-22.5l-50.1,54.5l-50.1,-54.5l-70.5,22.5l-16.3,-72.2l-72.4,-15.6l22,-70.6l-54.6,-49.9l54.6,-49.9l-22,-70.6l72.3,-15.6l16.3,-72.2l70.5,22.5l50.1,-54.5l50.1,54.5l70.5,-22.5l16.3,72.2l72.3,15.6l-22,70.6z" />
+    <path
+        android:fillColor="#FDC567"
+        android:pathData="M241.8,244.5m-138.7,0a138.7,138.7 0,1 1,277.5 0a138.7,138.7 0,1 1,-277.5 0" />
 </vector>
