+++ app/src/main/res/layout-land/activity_about.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/content_layout"
@@ -18,7 +17,7 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/img_sun"
-        tools:ignore="ContentDescription"/>
+        tools:ignore="ContentDescription" />
 
     <TextView
         android:id="@+id/app_name_text"
@@ -34,7 +33,7 @@
         app:layout_constraintLeft_toRightOf="@+id/icon"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="@+id/icon"
-        app:layout_constraintWidth_default="wrap"/>
+        app:layout_constraintWidth_default="wrap" />
 
     <TextView
         android:id="@+id/version_text"
@@ -49,7 +48,7 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/app_name_text"
         app:layout_constraintWidth_default="wrap"
-        tools:text="Version 1.0"/>
+        tools:text="Version 1.0" />
 
     <TextView
         android:id="@+id/author_text"
@@ -64,7 +63,7 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/version_text"
         app:layout_constraintVertical_bias="1.0"
-        app:layout_constraintWidth_default="wrap"/>
+        app:layout_constraintWidth_default="wrap" />
 
     <TextView
         android:id="@+id/images_license_text"
@@ -79,5 +78,5 @@
         app:layout_constraintHeight_default="wrap"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintWidth_default="wrap"/>
+        app:layout_constraintWidth_default="wrap" />
 </android.support.constraint.ConstraintLayout>
