+++ app/src/main/res/layout/fragment_weather.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.SwipeRefreshLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.v4.widget.SwipeRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/swipe_refresh_layout"
@@ -26,7 +25,7 @@
             android:textStyle="bold"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            tools:text="12 July 2017"/>
+            tools:text="12 July 2017" />
 
         <TextView
             android:id="@+id/temperature_text"
@@ -39,7 +38,7 @@
             android:textSize="88sp"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/date_text"
-            tools:text="123"/>
+            tools:text="123" />
 
         <TextView
             android:id="@+id/units_text"
@@ -50,7 +49,7 @@
             android:textSize="32sp"
             app:layout_constraintBaseline_toBaselineOf="@+id/temperature_text"
             app:layout_constraintLeft_toRightOf="@+id/temperature_text"
-            tools:text="ºC"/>
+            tools:text="ºC" />
 
         <TextView
             android:id="@+id/location_text"
@@ -64,7 +63,7 @@
             android:textSize="16sp"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/temperature_text"
-            tools:text="Moscow region, Dolgoprudny"/>
+            tools:text="Moscow region, Dolgoprudny" />
 
         <ImageView
             android:id="@+id/weather_image"
@@ -83,7 +82,7 @@
             app:layout_constraintTop_toBottomOf="@+id/date_text"
             app:layout_constraintWidth_default="spread"
             tools:ignore="ContentDescription"
-            tools:src="@drawable/img_sun"/>
+            tools:src="@drawable/img_sun" />
 
         <xyz.matteobattilana.library.WeatherView
             android:id="@+id/weather_view"
@@ -92,6 +91,6 @@
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/location_text"
-            app:layout_constraintWidth_default="spread"/>
+            app:layout_constraintWidth_default="spread" />
     </android.support.constraint.ConstraintLayout>
 </android.support.v4.widget.SwipeRefreshLayout>
