+++ app/src/main/res/layout/nav_header.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="@dimen/nav_header_height"
     android:background="@color/colorNavHeader"
@@ -16,7 +15,7 @@
         android:layout_height="wrap_content"
         android:text="@string/app_name"
         android:textColor="@color/colorNavHeaderText"
-        android:textSize="22sp"/>
+        android:textSize="22sp" />
 
     <android.support.design.widget.TextInputLayout
         android:id="@+id/input_layout_name"
