+++ app/src/main/res/menu/menu_nav_drawer.xml
@@ -4,9 +4,9 @@
     <item
         android:id="@+id/nav_settings"
         android:icon="@drawable/ic_settings"
-        android:title="@string/settings"/>
+        android:title="@string/settings" />
     <item
         android:id="@+id/nav_about"
         android:icon="@drawable/ic_help"
-        android:title="@string/about"/>
+        android:title="@string/about" />
 </menu>
