+++ build.gradle
@@ -19,7 +19,7 @@ allprojects {
         jcenter()
         mavenCentral()
         maven { url = 'https://jitpack.io' }
-        maven { url 'https://maven.google.com'}
+        maven { url 'https://maven.google.com' }
     }
 }
