+++ data/src/main/java/com/acbelter/weatherapp/data/netmodel/Main.java
@@ -1,4 +1,3 @@
-
 package com.acbelter.weatherapp.data.netmodel;
 
 import com.google.gson.annotations.Expose;
