+++ data/src/main/java/com/acbelter/weatherapp/data/netmodel/Wind.java
@@ -1,4 +1,3 @@
-
 package com.acbelter.weatherapp.data.netmodel;
 
 import com.google.gson.annotations.Expose;
