+++ data/src/main/java/com/acbelter/weatherapp/data/network/Api.java
@@ -4,11 +4,12 @@ import io.reactivex.Observable;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
+import static com.acbelter.weatherapp.data.BuildConfig.OPEN_WEATHER_MAP_API_KEY;
+
 public interface Api {
-    String OPEN_WEATHER_API_KEY = "5e09f6cd0f1eb1c1b5405f1327a62f7c";
     String BASE_URL = "http://api.openweathermap.org/data/2.5/";
 
-    @GET("weather?APPID=" + OPEN_WEATHER_API_KEY)
+    @GET("weather?APPID=" + OPEN_WEATHER_MAP_API_KEY)
     Observable<String> getCurrentWeatherData(@Query("q") String city,
                                              @Query("lang") String lang);
 }
