--- data/src/main/java/com/acbelter/weatherapp/data/network/ApiKeys.java
@@ -1,7 +1,0 @@
-package com.acbelter.weatherapp.data.network;
-
-public final class ApiKeys {
-
-    public static final String OPEN_WEATHER_API_KEY = "5e09f6cd0f1eb1c1b5405f1327a62f7c";
-    public static final String GOOGLE_PLACES_API_KEY = "AIzaSyBl3ZETwhkVGM_0_vXJrhi3hgJ37HT88X0";
-}
