+++ data/src/main/java/com/acbelter/weatherapp/data/network/LocationApi.java
@@ -6,7 +6,7 @@ import io.reactivex.Observable;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
-import static com.acbelter.weatherapp.data.network.ApiKeys.GOOGLE_PLACES_API_KEY;
+import static com.acbelter.weatherapp.data.BuildConfig.GOOGLE_PLACES_API_KEY;
 
 public interface LocationApi {
