+++ data/src/main/java/com/acbelter/weatherapp/data/placesmodel/Term.java
@@ -1,4 +1,3 @@
-
 package com.acbelter.weatherapp.data.placesmodel;
 
 import com.google.gson.annotations.Expose;
