+++ domain/src/main/AndroidManifest.xml
@@ -1,4 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest
-    package="com.acbelter.weatherapp.domain">
-</manifest>
+<manifest package="com.acbelter.weatherapp.domain"></manifest>
